To update tweet timestamps periodically
Tweet timestamps displayed in the short form relative to the current
time are updated periodically. The interval between updates is 57
seconds. The value of timestamp of each tweet is stored using jQuery’s
data method with the “time” key, i.e. $(tweetElement).data(“time”,
timestamp). The title attribute showing the date and time in the
original Crowy format is added to each timestamp link.